django/mikes_db_api/requirements.txt
apt-xapian-index==0.49
asgiref==3.4.1
attrs==20.3.0
blinker==1.4
cached-property==1.5.2
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
cryptography==3.3.2
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.5.0
distro-info==1.0
Django==4.0.1
django-cors-headers==3.10.1
djangorestframework==3.13.1
docker==4.4.4
docker-compose==1.27.4
dockerpty==0.4.1
docopt==0.6.2
galternatives==1.0.8
gyp==0.1
httplib2==0.18.1
idna==2.10
importlib-metadata==4.0.1
jeepney==0.7.1
jsonschema==3.2.0
keyring==23.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lubuntu-update-notifier==0.4.dev0
more-itertools==4.2.0
netifaces==0.10.9
oauthlib==3.1.0
olefile==0.46
pexpect==4.8.0
Pillow==8.1.2
psutil==5.8.0
psycopg2==2.9.3
pycairo==1.16.2
pycups==2.0.1
PyGObject==3.40.1
PyJWT==1.7.1
PyQt5==5.15.4
PyQt5-sip==12.9.0
pyrsistent==0.15.5
python-apt==2.2.1
python-debian===0.1.39ubuntu1
python-dotenv==0.15.0
pytz==2021.3
pyxdg==0.27
PyYAML==5.3.1
reportlab==3.5.66
requests==2.25.1
SecretStorage==3.3.1
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.2
systemd-python==234
terminator==2.1.0
texttable==1.6.3
ubuntu-advantage-tools==27.4
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.5
websocket-client==0.57.0
xdg==5
xkit==0.0.0
zipp==1.0.0
